fixed a few bugs

+++ main.cpp
@@ -11,7 +11,7 @@
 
 using namespace std;
 
-# define PATH "/home/blackreaper/Documents/Shell/"
+# define PATH "/home/hemant/Desktop/shell/"
 
 
 char poss_commands[18][50] = {
@@ -35,16 +35,6 @@ char poss_commands[18][50] = {
 	"calculator"
 };
 
-void getMatches(char *s){
-	int i = 0;
-	for (i = 0; i < sizeof(poss_commands) / 50; i++)
-	{
-		if (((string)poss_commands[i]).find((string)s) != string::npos)
-		{
-			printf("%s ", poss_commands[i]);
-		}
-	}
-}
 
 void execute_exit()
 {
@@ -70,7 +60,7 @@ int execute_cd(char **argv){
 		if (chdir(argv[1]) != 0)
 		{
 			perror("lsh");
-			return 0;
+			return 1;
 		}
 		else
 		{
@@ -131,76 +121,33 @@ char *read_input()
 	}
 
 	char c;
-	system("stty -icanon min 1");
 	while (1)
 	{
 		c = getchar();
 
-		if (c == EOF)
-		{
+		if(c == EOF) {
 			exit(EXIT_SUCCESS);
-		}
-		else if (c == '\n')
-		{
-			buffer[position] = '\0';
-			break;
-		}
-		else if (c == 0x7F)
-		{
-			position--;
-			if (position >= 0)
-			{
-				buffer[position] = '\0';
-				position--;
-			}
-		}
-		else if (c == '\t')
-		{
-			cout << "\n"
-				 << buffer << "\n";
-			printf("\n");
-			getMatches(buffer);
-			printf("\n");
-			position = 0;
-			position--;
-		}
-		else
-		{
-			buffer[position] = c;
-		}
-		position++;
+		} else if (c == '\n') {
+	    	buffer[position] = '\0';
+	    	return buffer;
+	    } else {
+	    	buffer[position] = c;
+	    }
+	    position++;
 
-		if (position >= buff_size)
-		{
-			buff_size += BUFFER_SIZE;
-			buffer = (char *)realloc(buffer, buff_size);
-			if (!buffer)
-			{
-				fprintf(stderr, "hha@LBP-shell: allocation error\n");
-				exit(EXIT_FAILURE);
-			}
-		}
+	    if (position >= buff_size) {
+		    buff_size += BUFFER_SIZE;
+		    buffer = (char*) realloc(buffer, buff_size);
+		    if (!buffer) {
+			    fprintf(stderr, "hha@LBP-shell: allocation error\n");
+			    exit(EXIT_FAILURE);
+		    }
+	    }
 	}
 
-	system("stty cooked");
 	return buffer;
 }
 
-int execute_run(char **argv);
-
-
-int execute_run(char **argv)
-{
-
-	FILE *fp = fopen(argv[1], "r");
-	if (fp == NULL)
-	{
-		printf("Error opening file!\n");
-		return 0;
-	}
-
-	return 1;
-}
 
 void loop_input(){
 	char *input_line;
@@ -257,40 +204,42 @@ void loop_input(){
 			    }
 			}
 
-			if (strcmp(splitted_input[0], "cd") == 0) {
-				status = execute_cd(splitted_input);
-				matched = true;
-			}
-			else if (strcmp(splitted_input[0], "pwd") == 0) {
-				status = execute_pwd(splitted_input);
-				matched = true;
-			}
-			else if (strcmp(splitted_input[0], "exit") == 0) {
-				free(input_line);
-				free(splitted_input);
-				history_file.close();
-				execute_exit();
-			}
-			else{
-				int pid = fork();
-				if(pid < 0) fprintf(stderr, "fork error!");
-				else if(pid == 0){
-					char current_path[1024];
-					getcwd(current_path, sizeof(current_path));
-					strcat(current_path, "/");
-					strcat(current_path, splitted_input[0]);
-					int executed = execvp(current_path, splitted_input);
-					if(executed == -1) {
-						printf("The command entered is not supported by this shell!!!\n");
-						kill(getpid(), SIGKILL);
-					}
+			if(!matched){
+				if (strcmp(splitted_input[0], "cd") == 0) {
+					status = execute_cd(splitted_input);
+					matched = true;
+				}
+				else if (strcmp(splitted_input[0], "pwd") == 0) {
+					status = execute_pwd(splitted_input);
+					matched = true;
+				}
+				else if (strcmp(splitted_input[0], "exit") == 0) {
+					free(input_line);
+					free(splitted_input);
+					history_file.close();
+					execute_exit();
 				}
 				else{
-					if ((pid = waitpid(pid, &status, 0)) < 0)
-						fprintf(stderr, "pid error!");
-					status = 1;
+					int pid = fork();
+					if(pid < 0) fprintf(stderr, "fork error!");
+					else if(pid == 0){
+						char current_path[1024];
+						getcwd(current_path, sizeof(current_path));
+						strcat(current_path, "/");
+						strcat(current_path, splitted_input[0]);
+						int executed = execvp(current_path, splitted_input);
+						if(executed == -1) {
+							printf("The command entered is not supported by this shell!!!\n");
+							kill(getpid(), SIGKILL);
+						}
+					}
+					else{
+						if ((pid = waitpid(pid, &status, 0)) < 0)
+							fprintf(stderr, "pid error!");
+						status = 1;
+					}
+					matched = true;
 				}
-				matched = true;
 			}
 
 			if(!matched){
